% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anony.R
\name{anony}
\alias{anony}
\title{Anonymize a text/factor variable by replacing unique values with random words}
\usage{
anony(data, var_name, word_list = pokemonz, seed = NULL)
}
\arguments{
\item{data}{A dataframe containing the variable to anonymize}

\item{var_name}{The name of the variable to anonymize}

\item{word_list}{A character vector of words to use for anonymization}

\item{seed}{An optional seed for reproducible randomization}
}
\value{
The dataframe with the specified variable anonymized
}
\description{
Anonymize a text/factor variable by replacing unique values with random words
}
\examples{
# Using palmerpenguins dataset
library(palmerpenguins)
anonymized_penguins <- anony(penguins, species, pokemonz, 123)

}
