% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_tools.R
\docType{data}
\name{cbfriendly}
\alias{cbfriendly}
\title{Color-blind friendly palette}
\format{
A character vector with 8 elements.
}
\usage{
cbfriendly
}
\description{
A list of color-blind friendly distinct colors
}
\examples{
cbfriendly
pie(rep(1,8), col=cbfriendly, labels=cbfriendly)
}
\keyword{datasets}
