% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/small_functions.R
\name{clip2csv}
\alias{clip2csv}
\title{Clip to CSV}
\usage{
clip2csv(x, row.names = FALSE, col.names = TRUE, ...)
}
\arguments{
\item{x}{Dataframe to copy}

\item{row.names}{(optional) logical}

\item{col.names}{(optional) logical}
}
\value{
Pastable table in comma delimited text format
}
\description{
Collects a dataframe as clipboard contents.
}
\details{
Do not use for massive dataframes

Example usage: clip2csv(df)
}
\examples{
clip2csv(mtcars)
}
