% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/small_functions.R
\name{clip2xl}
\alias{clip2xl}
\title{Clip to Excel}
\usage{
clip2xl(x, row.names = FALSE, col.names = TRUE, ...)
}
\arguments{
\item{x}{Dataframe to copy}

\item{row.names}{(optional) logical}

\item{col.names}{(optional) logical}
}
\value{
Pastable table for excel
}
\description{
Collects a dataframe as clipboard contents.
}
\details{
Do not use for massive dataframes

Example usage: clip2xl(df)
}
\examples{
clip2xl(iris)
}
