% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/small_functions.R
\name{clipfromxl}
\alias{clipfromxl}
\title{Clip from Excel}
\usage{
clipfromxl()
}
\description{
Make a dataframe with clipboard contents, from Excel
}
\details{
Should include header row, tidy column names will be produced

Example usage: df <- clipfromxl()
}
\examples{
df <- clipfromxl()
}
