% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coselectr.R
\name{coselectr}
\alias{coselectr}
\title{Coselectr Interactive Column Selector for DataFrames}
\usage{
coselectr(df)
}
\arguments{
\item{df}{A data frame to explore}
}
\value{
Selected columns from the input data frame
}
\description{
Quickly select columns of interest from large dataframes
}
\details{
Deselect by unchecking columns. Columns can be previewed in the table, 
and list at the top can be copied for pasting into console or R script for 
use in selecting the chosen columns
}
\examples{
coselectr(mtcars)
selected_data <- coselectr(mtcars)
}
