% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/small_functions.R
\name{\%nin\%}
\alias{\%nin\%}
\title{Not in}
\usage{
... \%nin\% NA
}
\value{
Logical list
}
\description{
Easier negative filtering
}
\details{
Example usage: aliens <- starwars %>% filter(species %nin% c("Human", "Droid"))
}
\examples{
iris_filtered <- subset(iris, (Species \%nin\% c("setosa", "versicolor")))
}
