% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opera_utils.R
\name{harmread_ob}
\alias{harmread_ob}
\title{Harmony Object Reader}
\usage{
harmread_ob(pthname, flename)
}
\arguments{
\item{pthname}{Path or directory to read}

\item{flename}{Filename pattern}
}
\value{
Dataframe
}
\description{
Imports all Harmony object results from directory and subdirectories.
}
\details{
Any file which matches the filename specified is imported to one dataframe.
Platename and Evaluation number is added to each row,
this info comes from the metadata in the head of the txt file.
Multiple plates are appended to the bottom of the resulting dataframe
}
\examples{
df <- harmread_ob("/my_subdirectory", "Objects_Population - fov.txt")
pth <- getwd()
df <-harmread_pl(pth, "Objects_Population - nuclei.txt")
}
