% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opera_utils.R
\name{harmread_pl}
\alias{harmread_pl}
\title{Harmony Reader}
\usage{
harmread_pl(pthname)
}
\arguments{
\item{pthname}{Path or directory to read}
}
\value{
Dataframe
}
\description{
Imports all Harmony plate results from directory and subdirectories.
}
\details{
Any file named using Harmony default convention of PlateResults.txt is included
Platename and Evaluation number is added to each row,
this info comes from the metadata in the head of the txt file.
Multiple plates are appended to the bottom of the resulting dataframe
}
\examples{
df <- harmread_pl("/my_subdirectory")
pth <- getwd()
df <-harmread_pl(pth)
}
