% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opera_utils.R
\name{hotplate}
\alias{hotplate}
\title{Hotplate}
\usage{
hotplate(dfname, metricname)
}
\arguments{
\item{dfname}{Dataframe}

\item{metricname}{Variable to heatmap}
}
\value{
ggplot object
}
\description{
Makes a heatmap of the plate
}
\details{
Needs numerical row and column info
only works for 96 well plate at present 
if numerical row col not available, run well2row first
}
\examples{
hotplate(df, nuclei_number_of_objects)
}
