% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/small_functions.R
\name{jiggle}
\alias{jiggle}
\title{Jiggle}
\usage{
jiggle(dfname)
}
\arguments{
\item{dfname}{Dataframe}
}
\value{
Re-ordered dataframe
}
\description{
Relocate (aka reorder) columns in a dataframe so that character columns are first.
}
\details{
Example usage: df <- jiggle(df)
}
\examples{
iris <- jiggle(iris)
}
