% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loggy.R
\name{loggy}
\alias{loggy}
\title{Loggy Log transformation for numeric variables}
\usage{
loggy(df, base = exp(1), offset = 0)
}
\arguments{
\item{df}{A data frame}

\item{base}{Base for logging, default is e ≈ 2.718}

\item{offset}{To make all columns positive, default is 0}
}
\value{
Data frame with numeric columns log_transformed
}
\description{
One function to log transform all numeric variables in a dataset
}
\details{
All newly created columns names will be appended with log_
Only numeric columns are transformed, default log is natural log, but other bases
e.g. 2 or 10 can be specified.   
The offset is calculated automatically to prevent log of negative numbers but
can be added manually.
}
\examples{
loggy(mtcars)
loggy(mtcars, 10)
}
