% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_tools.R
\name{majsty}
\alias{majsty}
\title{Majsty interactive scatter plots}
\usage{
majsty(data, x_axis_variable = NULL, y_axis_variable = NULL)
}
\arguments{
\item{data}{Dataframe to plot}

\item{x_axis_variable}{Data to plot on x axis (optional, must be in quotes)}

\item{y_axis_variable}{Data to plot on y axis (optional, must be in quotes)}
}
\value{
Interactive scatterplot
}
\description{
Select and colour points in a scatterplot
}
\details{
A shiny application to create scatter plots with selectable x and y axes for EDA

Points can be highlighted in different colours, and importantly these colours persist
when the axes are changed, for example to visualise where outliers in one metric
appear in other metrics.  
Starting X and Y variables can be preselected
}
\examples{
majsty(df)
majsty(user_df, "no_of_cells", "stain_intensity")
}
