% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opera_utils.R
\name{platemap}
\alias{platemap}
\title{Platemap}
\usage{
platemap(dfname, metricname = compound)
}
\arguments{
\item{dfname}{Dataframe}

\item{metricname}{Sample type, default is compound}
}
\value{
ggplot object
}
\description{
Makes a colour map of samples on the plate.
}
\details{
Requires row and column in numerical format
Only works for 96 well at present. Dataframes with duplicate or multiple 
values for each well may cause problems, so aim for 96 or less observations.
}
\examples{
platemap(df, compound)
}
