% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_tools.R
\name{ploxy}
\alias{ploxy}
\title{Ploxy interactive scatter plots}
\usage{
ploxy(data)
}
\arguments{
\item{data}{Dataframe to plot}
}
\value{
Interactive scatterplot
}
\description{
Shiny and plotly combine for scatterplot interaction
}
\details{
A shiny application to create scatter plots with selectable x and y axes for EDA.
X and Y variables can be interactively changed, along with
Plotly actions such as zoom, pan and save.
}
\examples{
ploxy(df)
ploxy(mtcars)
}
