% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anony.R
\docType{data}
\name{pokemonz}
\alias{pokemonz}
\title{Pokemon name list}
\format{
A character vector with 148 elements.
}
\usage{
pokemonz
}
\description{
A List of Single word pokemon
}
\details{
This vector contains a list of 148 distinct Pokemon names; 
designed for use with the [anony()] function
}
\examples{
pokemonz
anonymized_penguins <- anony(penguins, "species", pokemonz)
}
\keyword{datasets}
