% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prism_reformat.R
\name{prismy_col}
\alias{prismy_col}
\title{Reformat for Prism Columns}
\usage{
prismy_col(ddframe, met4prism, grp)
}
\arguments{
\item{ddframe}{Dataframe}

\item{met4prism}{Column to take values from}

\item{grp}{Metadata column to use as column names in output}
}
\value{
Dataframe
}
\description{
Reformats your data so each group specified has its own column in prism
}
\details{
This is for a single column per group; replicates are listed vertically
}
\examples{
forprism <- prismy_col(df, number_of_cells, compound)
xx <- starwars \%>\% filter(!is.na(species) & !is.na(height))
prism_out <- prismy_col(xx, height, species)
}
