% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prism_reformat.R
\name{prismy_time}
\alias{prismy_time}
\title{Reformat for Prism by timepoint}
\usage{
prismy_time(
  ddframe,
  met4prism,
  grp,
  timemeasure = timepoint,
  wellid = metadata_well
)
}
\arguments{
\item{ddframe}{Dataframe}

\item{met4prism}{Column to take values from}

\item{grp}{Text column to use as column names in output}

\item{timemeasure}{Column to use as times in output, defaults to timepoint}

\item{wellid}{Well name to group replicates by, defaults to metadata_well}
}
\value{
Dataframe
}
\description{
Reformats your data so each group (e.g. compound) has its own column
and each replicate is a subcolumn. 
Timepoints are arranged vertically.
}
\details{
This makes an output suitable for pasting into Prisms grouped table format.
}
\examples{
for_prism <- prismy_time(example_timecourse, intensity, compound)
prism_out <- prismy_time(example_timecourse, intensity, compound, timepoint, metadata_well)
}
