% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/small_functions.R
\name{repspace}
\alias{repspace}
\title{Replace underscore with space}
\usage{
repspace(x, y = "_")
}
\arguments{
\item{x}{Variable you wish to remove underscores from}
}
\value{
A variable with spaces instead of underscores
}
\description{
Name says it all.
}
\details{
This is the reverse of repunderscore.
Example usage: user_df$new <- repspace(user_df$colname)
}
\examples{
mtcars$new <- repspace(rownames(mtcars))
mylist <- c("green_apple", "red_banana")
mylist <- repspace(mylist)
}
