% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/small_functions.R
\name{repunderscore}
\alias{repunderscore}
\title{Replace space with underscore}
\usage{
repunderscore(x, y = " ")
}
\arguments{
\item{x}{Variable you wish to remove spaces from}
}
\value{
A variable with underscores replacing any whitespaces
}
\description{
Self explanatory
}
\details{
Example usage: mtcars$newname <- repunderscore(rownames(mtcars))
}
\examples{
mtcars$new <- repspace(rownames(mtcars))
mylist <- c("green apple", "red banana")
mylist <- repspace(mylist)
}
