% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/small_functions.R
\name{rescale_z}
\alias{rescale_z}
\title{ZScore a variable}
\usage{
rescale_z(x)
}
\arguments{
\item{x}{Numeric variable you wish to rescale}
}
\value{
A new variable scaled by mean and sd
}
\description{
via Hadley Wickham/Twitter
}
\details{
Example usage: user_df$new <- rescale_z(user_df$colname)
}
\examples{
mtcars$z_mpg <- rescale_z(mtcars$mpg)
}
