% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/small_functions.R
\name{squish}
\alias{squish}
\title{Squish a variable}
\usage{
squish(x, min, max)
}
\arguments{
\item{x}{Numeric variable you wish to limit}

\item{min}{Lower limit for variable}

\item{max}{Upper limit for variable}
}
\value{
A new variable confined by upper and lower bounds
}
\description{
Specifies lower and upper bounds of a variable to defined values
via Hadley Wickham/Twitter
Any values below minimum/ above maximum will be replaced with min/max value respectively
Example usage: user_df$new <- squish(user_df$colname)
}
\examples{
mtcars$limits <- squish(mtcars$hp, 100, 220)
}
