% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_tools.R
\name{theme_js}
\alias{theme_js}
\title{James Theme}
\usage{
theme_js()
}
\description{
A minimal ggplot theme
}
\details{
This is superseded by theme_tt but kept for continuity with old scripts
}
\examples{
ggplot(data = mtcars, aes(x = wt, y = mpg, colour = factor(cyl))) +
  geom_point() +
  ggtitle("JS theme") +
  theme_js()
}
