% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_tools.R
\name{theme_tt}
\alias{theme_tt}
\title{Talisman Theme}
\usage{
theme_tt()
}
\description{
A minimal ggplot theme.
}
\details{
Talisman colours not yet included.  

Showtext package needs to be installed and Rstudio Global Options...General...Graphics...Backend... 
should be set to Cairo PNG (or AGG, basically anything except "windows").  
If this errors when knitting, add library(extrafont) to Rmd

This package requires font "Fira Sans" to be installed on your system - download from google fonts and restart R if not displaying correctly.
}
\examples{
ggplot(data = mtcars, aes(x = wt, y = mpg, colour = factor(cyl))) +
  geom_point() +
  ggtitle("Talisman theme") +
  theme_tt()
}
