% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_tools.R
\name{theme_tt2}
\alias{theme_tt2}
\title{Talisman Theme 2}
\usage{
theme_tt2()
}
\description{
A minimal ggplot theme with Fira Sans font which must be installed on your system, via google fonts.
}
\details{
This requires the following in the header of your Qmd file: 
knitr:
  opts_chunk:
    dev: ragg_png

A minimal theme for ggplot with legend at the bottom and no gridlines.
}
\examples{
ggplot(data = mtcars, aes(x = wt, y = mpg, colour = factor(cyl))) +
  geom_point() +
  ggtitle("Talisman theme") +
  theme_tt2()
}
