% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_tools.R
\name{theme_tt2f}
\alias{theme_tt2f}
\title{Talisman Theme for Facets 2}
\usage{
theme_tt2f()
}
\description{
A minimal ggplot theme for faceted charts, using Fira Sans.
}
\examples{
ggplot(data = mtcars, aes(x = wt, y = mpg, colour = factor(cyl))) +
  geom_point() +
  ggtitle("Talisman theme") +
  facet_wrap(~cyl) +
  theme_tt2f()
}
