% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_tools.R
\name{theme_ttf}
\alias{theme_ttf}
\title{Talisman Theme for Facets}
\usage{
theme_ttf()
}
\description{
A minimal ggplot theme for faceted charts.
}
\details{
Rstudio Global Options...General...Graphics...Backend... 
       should be set to Cairo PNG (or anything except "windows").
}
\examples{
ggplot(data = mtcars, aes(x = wt, y = mpg, colour = factor(cyl))) +
  geom_point() +
  ggtitle("Talisman theme") +
  facet_wrap(~cyl) +
  theme_ttf()
}
