% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_tools.R
\docType{data}
\name{tubemap}
\alias{tubemap}
\title{Tubemap colour list}
\format{
A character vector with 23 elements.
}
\usage{
tubemap
}
\description{
A List of colours inspired by the London Underground Map
}
\details{
This vector contains a list of distinct 23 colours: after 16 it becomes greyscale.
Note that for optimal visual distinction, it's recommended to use no more than 16 colors.
}
\examples{
tubemap
pie(rep(1,23), col=tubemap, labels=tubemap)
}
\keyword{datasets}
