% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/small_functions.R
\name{u}
\alias{u}
\title{Unique time saver}
\usage{
u(my_var, dfname = df)
}
\arguments{
\item{my_var}{Variable to extract unique values}

\item{dfname}{Dataframe}
}
\value{
Re-ordered dataframe
}
\description{
A shortcut for the unique function
}
\details{
When working with a new data set, the unique() function is often used
multiple times to get a short list of info within the data such as 
treatment groups or plate barcodes. This is a function for the lazy typist
to reduce unique(df$my_var) to u(my_var)

Eight keystrokes saved every time!

If your dataframe is not df, there is an option to specify dfname
}
\examples{
df <- iris
u(Species)
u(cyl, mtcars)
}
