% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opera_utils.R
\name{well2row}
\alias{well2row}
\title{Well2row}
\usage{
well2row(dfname, wellvar)
}
\arguments{
\item{dfname}{Dataframe}

\item{wellvar}{Variable to convert}
}
\value{
Dataframe
}
\description{
Makes row and column numbers
Requires well name in A01 format
}
\examples{
new_df <- well2row(df, well)
}
