% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_tools.R
\name{barchartr}
\alias{barchartr}
\title{Quick summary barcharts}
\usage{
barchartr(dfname, varname, groupnames)
}
\arguments{
\item{varname}{Variable to plot on barchart}

\item{groupnames}{Variable which lists groups which will become bar labels}

\item{df}{Dataframe to plot}
}
\value{
ggplot object
}
\description{
Plots mean with SD error bars in GGplot
}
\details{
A quick function to plot summary data barcharts with error bars calculated means.

A time-saver since GGplot doesn't do this by default, but customisation options are limited.  
Default colours are based on London underground map. Recommended maximum number of bars is 16, absolute limit is 23.
}
\examples{
barchartr(df, number_of_cells, compound)
bc <- barchartr(df, GFP_pos_cells, treatment)
}
