% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_tools.R
\name{ploxytime}
\alias{ploxytime}
\title{Ploxytime interactive scatter plots with animation}
\usage{
ploxytime(data, timevariable = "timepoint")
}
\arguments{
\item{data}{Dataframe to plot}

\item{timevariable}{Name of the column to use for animation frames (default: "timepoint")}
}
\value{
Interactive scatterplot with animation
}
\description{
Shiny and plotly combine for timecourse interaction
}
\details{
A shiny application to create scatter plots with selectable x and y axes for EDA.
X and Y variables can be interactively changed, along with
Plotly actions such as zoom, pan and save.
}
\examples{
ploxytime(df)
ploxytime(df, "date")
}
