% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_tools.R
\name{theme_tt}
\alias{theme_tt}
\title{Talisman Theme}
\usage{
theme_tt()
}
\description{
A minimal ggplot theme.
}
\details{
Talisman colours not yet included.  

A minimal theme for ggplot with legend at the bottom and no gridlines.
}
\examples{
ggplot(data = mtcars, aes(x = wt, y = mpg, colour = factor(cyl))) +
  geom_point() +
  ggtitle("Talisman theme") +
  theme_tt()
}
